Ensure the same make arguments are used for install.
Without this we had two issues.  Firstly, the entire build was restarted
during the "install" phase because the command line is different, so we
end up building the whole thing twice, and to make things worse the
second time was significantly slower as install is invoked without
parallel jobs.

Secondly, and more importantly, because the second build did not pass
CFLAGS and CXXFLAGS, they were never used for the final binaries!