Add support for building current nodes on 24.4.1.
Rather than trying to upgrade node at the same time, for now this just
builds the same versions as previously (v8.17.0 and v6.17.1).  However,
a couple of new patches are still required:

 * dtrace-binutils241-v8.x.patch to fix dtrace build with binutils-2.41.
 * v8-includes.patch to fix build with gcc13.

In addition I have added a few -Wno flags to cut down on the amount of
warnings spewed during the build.  Given these are very old versions of
node, there is no chance that they are going to be cleaned up, and this
just results in the logs being incredibly large (it's impossible to view
them in Jenkins for example, browsers just hang).