Build relocatable v18.20.6, v8.17.0, v6.17.1.
As per comments in TRITON-2428, this simplifies the build to only
produce what used to be called the "gz" builds, i.e. fully relocatable
builds that bundle the GCC runtime libraries.

The v6.17.1 build requires building with GCC 7 to avoid nasty C++
behaviour that leads to segfaults, and bundling the GCC libraries is
simpler and more robust than having to ship multiple gcc*-libs packages
in every image that require these builds.

The v6.x and v8.x builds retain patches required by TOOLS-1950 and
TRITON-1311 to modify the default behaviour for certain components,
while the new v18.20.6 build is unmodified to ensure that the test suite
is fully clean, and so that components can be rewritten to use the newer
APIs when they are migrated over.